requirements.txt
Django==3.1.2
django-cors-middleware==1.5.0
djangorestframework==3.12.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1

